refactor(terminal): PR #1220 write outcome責務を分離
