+++ src/types/terminal-write.ts
@@ -1,0 +1,10 @@
+export type TerminalWriteOutcome =
+  | 'written'
+  | 'suppressedInjecting'
+  | 'droppedTooLarge'
+  | 'droppedRateLimited'
+  | 'sessionNotFound';
+
+export interface TerminalWriteResult {
+  outcome: TerminalWriteOutcome;
+}
